pyproject.toml
[project]
name = "psoriasis-robust-adv-defense"
version = "0.1.0"
description = "A Study on Collaborative Adversarial Attack-Defense Mechanisms for Robustness Enhancement in Psoriasis Medical Image Diagnosis."
authors = [
    { name = "WJ.Cao" },
    { name = "Q.Cheng" },
    { name = "YX.Li" },
    { name = "ZY.Feng" },
    { name = "YX.Li" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.35.3",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scipy>=1.17.1",
    "torch==2.9.0",
    "torchvision==0.24.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["psorad*"]

# 定义额外的 PyTorch CUDA 索引
[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

# 指定 torch 使用 PyTorch 专用索引
[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }

# ruff
[tool.ruff]
line-length = 240
target-version = "py312"
exclude = [".venv", "__pycache__", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "N"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]

[project.scripts]
psorad = "psorad.cli:main"
